pyproject.toml
[project]
name = "haratch-ocr"
version = "0.1.0"
description = "OCR processor for the Armenian newspaper Haratch"
authors = [
    {name = "Romain Florentz"}
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch>=2.7.1,<3.0.0",
    "torchvision>=0.22.1,<0.23.0",
    "numpy>=2.2.6,<3.0.0",
    "doclayout-yolo>=0.0.3,<0.0.4",
    "fire>=0.7.0,<0.8.0",
    "huggingface-hub>=0.32.4,<0.33.0",
    "pytesseract>=0.3.13,<0.4.0",
    "pdf2image>=1.17.0,<2.0.0",
    "pypdf2>=3.0.1,<4.0.0",
    "requests>=2.31.0,<3.0.0",
    "google-cloud-storage>=3.7.0",
]

[project.optional-dependencies]
translate = [
    "python-dotenv>=1.1.0,<2.0.0",
    "openai>=1.84.0,<2.0.0",
    "google-generativeai>=0.8.5,<0.9.0",
]
